fix(integrator): widen counter members to int64_t
Replaces `int` with `int64_t` for `naccept_`, `nreject_`, and
`n_failed_event_refinements_` so that `BatchCounterWriteback` /
`BatchEventCounterWriteback` aggregation across thousands of
trajectories × millions of steps cannot silently wrap under NDEBUG.
The prior assert-only narrowing guard was the only defense and was
absent in Release; widening removes the failure mode entirely.

Per-trajectory `int` locals threaded through driver loops are
unchanged — they remain bounded by `max_steps` (also `int`). The
existing `static_cast<long long>` cap-checks in adaptive_driver.h
and parallel_driver.h are kept as defensive widening for the
`naccept + nreject` sum.

Test files that captured `get_naccept`/`get_nreject` into `int`
locals are widened to `int64_t` (no behavior change).

+++ tests/cpp/integrators/test_copy_settings.cpp
@@ -84,9 +84,9 @@ TEST(CopySettingsFromTest, AllFieldsRoundTripAcrossDifferentDODE) {
 
     // Snapshot observable fields that copy_settings_from MUST NOT touch,
     // so we can verify they're unchanged after the copy.
-    const int pre_naccept = tgt.get_naccept();
-    const int pre_nreject = tgt.get_nreject();
-    const int pre_failed_event = tgt.get_failed_event_count();
+    const int64_t pre_naccept = tgt.get_naccept();
+    const int64_t pre_nreject = tgt.get_nreject();
+    const int64_t pre_failed_event = tgt.get_failed_event_count();
 
     tgt.copy_settings_from(src);
